+++ server/lib/rsa/_version200.py
@@ -336,7 +336,7 @@ def extended_gcd(a, b):
     oa = a                             #Remember original a/b to remove
     ob = b                             #negative values from return results
     while b != 0:
-        q = long(a/b)
+        q = int(a/b)
         (a, b)  = (b, a % b)
         (x, lx) = ((lx - (q * x)),x)
         (y, ly) = ((ly - (q * y)),y)
@@ -397,7 +397,7 @@ def encrypt_int(message, ekey, n):
     """Encrypts a message using encryption key 'ekey', working modulo n"""
 
     if type(message) is IntType:
-        message = long(message)
+        message = int(message)
 
     if not type(message) is LongType:
         raise TypeError("You must pass a long or int")
